Initial Prototype
The App.js file that will take in a JSON file from the local path will then extract the needed data to create a webpage.

Included the used JSON file

change the file path in the code to your local path of where the JSON file is saved in order for the code to work.